requirements.txt
grpcio==1.44.0
grpcio-tools==1.44.0
pandas==1.3.5
protobuf==3.19.4
pyparsing==3.0.6
typing==3.7.4.3

